Move logic for initializing and drawing the chart/svg and the event listeners for the chart form headstart.js to new file canvas.js A lot of calls to headstart changed to calls to my_canvas
